fix: copy Whisper mobile benchmark with integration tests
